# 2113 using compact incomplete on a library with dynamic schema with a named index can result in an unreadable index