--- a/cpp/arcticdb/version/schema_checks.cpp
+++ b/cpp/arcticdb/version/schema_checks.cpp
@@ -75,8 +75,8 @@ bool index_names_match(
     const StreamDescriptor& df_in_store_descriptor,
     const StreamDescriptor& new_df_descriptor
 ) {
-    auto df_in_store_index_field_count = df_in_store_descriptor.index().type() == IndexDescriptor::Type::EMPTY ? 0 : df_in_store_descriptor.index().field_count();
-    auto new_df_field_index_count = new_df_descriptor.index().type() == IndexDescriptor::Type::EMPTY ? 0 : new_df_descriptor.index().field_count();
+    auto df_in_store_index_field_count = df_in_store_descriptor.index().field_count();
+    auto new_df_field_index_count = new_df_descriptor.index().field_count();
 
     // If either index is empty, we consider them to match
     if (df_in_store_index_field_count == 0 || new_df_field_index_count == 0) {