--- a/python/tests/hypothesis/arcticdb/test_append.py
+++ b/python/tests/hypothesis/arcticdb/test_append.py
@@ -6,8 +6,8 @@ from arcticdb.version_store import NativeVersionStore
 from arcticdb_ext.exceptions import (
     InternalException,
     NormalizationException,
-    StreamDescriptorMismatch,
 )
+from arcticdb_ext.version_store import StreamDescriptorMismatch
 from arcticdb_ext import set_config_int
 from hypothesis import given, assume, settings, strategies as st
 from itertools import chain, product, combinations