… python/tests/unit/arcticdb/version_store/test_parallel.py
… python/tests/unit/arcticdb/version_store/test_sort_merge.py