@G-D-Petrov 2025-01-14
The benchmarks are reporting ~30% performance degradation on the FinalizeStagedData benchmarks:

| Change   | Before [d71a0bb5] <v5.2.0rc0~1>   | After [c1b389ee]    |   Ratio | Benchmark (Parameter)                                                      |
|----------|-----------------------------------|---------------------|---------|----------------------------------------------------------------------------|
| +        | 904M                              | 1.45G               |    1.6  | finalize_staged_data.FinalizeStagedData.peakmem_finalize_staged_data(1000) |
| +        | 1.88G                             | 2.73G               |    1.45 | finalize_staged_data.FinalizeStagedData.peakmem_finalize_staged_data(2000) |
| +        | 1.71±0s                           | 2.33±0s             |    1.36 | finalize_staged_data.FinalizeStagedData.time_finalize_staged_data(1000)    |
| +        | 3.49±0s                           | 4.66±0s             |    1.34 | finalize_staged_data.FinalizeStagedData.time_finalize_staged_data(2000)    |

I think that this is due to [the new check over all of the segments to make sure that the index names are the same](https://github.com/man-group/ArcticDB/pull/2116/files#diff-101f018db937d9932cd415402f8ff60559552ce72b63ff2efd3536bd83d828deR1085), which was not done before.