@vasil-pashov 2025-01-15
I think it's worth adding a test for `sort_and_finalize_staged_data` similar to the one for `finalize_staged_data` and the codepaths are slightly different